+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Fixed
 - Show replies count for messages [#4844](https://github.com/rokwire/illinois-app/issues/4844).
+- Fixed Analytics logs for Event2 onboarding screen [#4501](https://github.com/rokwire/illinois-app/issues/4501).
 
 ## [6.1.68] - 2025-03-07
 ### Changed